requirements.txt
flask
transformers
torch
accelerate
huggingface_hub
python-dotenv